[AND-690] Extract list impression tracking to be able to be re-used
